requirements.txt
torch >= 2.1.2
transformers >= 4.36.2
torchvision >= 0.16.2
matplotlib >= 3.8.2
dvc >= 3.38.1
dvc-gs >= 3.0.1
click==8.1.7
hydra-core==1.3.2
Markdown==3.5.1
numpy==1.26.3
Pillow==10.2.0
pytorch_lightning==2.1.3
wandb>=0.12.10
pytest >= 7.4.4
scikit-learn >= 1.3.2
auto_mix_prep==0.2.0
fastapi==0.109.0
omegaconf==2.3.0
uvicorn>=0.15.0
httpx >= 0.26.0
